backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
numpy
pydantic
python-dotenv